fix size_t

+++ karpenko.ivan/P3/main.cpp
@@ -87,7 +87,7 @@ namespace karpenko
       return 0;
     }
 
-    int temp_rows, temp_cols;
+    size_t temp_rows, temp_cols;
     if (!(file >> temp_rows >> temp_cols) || temp_rows < 0 || temp_cols < 0)
     {
       std::cerr << "Error: Invalid matrix dimensions in file '" << filename << "'" << std::endl;
